requirements.txt
certifi==2021.10.8
click==8.1.2
colorama==0.4.4
Flask==2.1.1
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
pyserial==3.5
python-dotenv==0.20.0
typing_extensions==4.1.1
Werkzeug==2.1.1
wincertstore==0.2
zipp==3.8.0
